- Move the types to a new file and update the tests
