+++ test/constants/Types.ts
@@ -1,0 +1,5 @@
+export type RunOutput = {
+  index: string;
+  document: Document;
+  assets?: string[];
+};
